Set up nullmailer and add SMTP relay settings.
The various web apps will use nullmailer to send email on to the
configured email relay. Pass --quiet to the git fetch in the homepage
update script so that network outages do not send an email every 5
minutes. Also fix some typos in the homepage update script. Edit the
cron to not send emails for that update script either.